+++ _maps/RandomRuins/JungleRuins/jungle_cavecrew.dmm
@@ -1762,7 +1762,7 @@
 /turf/open/floor/plasteel/tech,
 /area/ruin/jungle/cavecrew/security)
 "vr" = (
-/obj/machinery/computer/cargo/express,
+/obj/machinery/computer/cargo,
 /turf/open/floor/plasteel/patterned,
 /area/ruin/jungle/cavecrew/cargo)
 "vH" = (
