+++ _maps/outpost/hangar/nt_asteroid_20x20.dmm
@@ -370,7 +370,7 @@
 	},
 /area/hangar)
 "jw" = (
-/obj/machinery/computer/cargo/express,
+/obj/machinery/computer/cargo,
 /obj/structure/railing{
 	dir = 8;
 	layer = 4.1
