+++ _maps/outpost/hangar/nt_asteroid_40x20.dmm
@@ -839,7 +839,7 @@
 	},
 /area/hangar)
 "tH" = (
-/obj/machinery/computer/cargo/express,
+/obj/machinery/computer/cargo,
 /obj/item/toy/plush/knight{
 	pixel_y = 25;
 	pixel_x = 9
