+++ _maps/outpost/hangar/nt_asteroid_40x40.dmm
@@ -60,7 +60,7 @@
 /obj/effect/turf_decal/techfloor{
 	dir = 4
 	},
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/machinery/light/directional/east,
