+++ _maps/outpost/hangar/nt_asteroid_56x20.dmm
@@ -284,7 +284,7 @@
 	},
 /area/hangar)
 "kx" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8;
 	pixel_x = 7
 	},
