+++ _maps/outpost/hangar/nt_asteroid_56x40.dmm
@@ -723,7 +723,7 @@
 	},
 /area/hangar)
 "Ed" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8;
 	pixel_x = 7
 	},
