+++ _maps/outpost/nanotrasen_asteroid.dmm
@@ -5585,7 +5585,7 @@
 /turf/open/floor/concrete/reinforced,
 /area/outpost/maintenance/aft)
 "tW" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/effect/turf_decal/techfloor{
