+++ _maps/shuttles/independent/independent_beluga.dmm
@@ -4742,7 +4742,7 @@
 /area/ship/crew)
 "UO" = (
 /obj/effect/turf_decal/industrial/traffic/corner,
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 1
 	},
 /obj/machinery/light/directional/south,
