+++ _maps/shuttles/independent/independent_box.dmm
@@ -1802,7 +1802,7 @@
 /obj/effect/turf_decal/corner/opaque/blue{
 	dir = 4
 	},
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 1
 	},
 /obj/machinery/firealarm/directional/south,
