+++ _maps/shuttles/independent/independent_bubble.dmm
@@ -1164,7 +1164,7 @@
 	},
 /obj/item/circuitboard/computer/selling_pad_control,
 /obj/item/circuitboard/machine/selling_pad,
-/obj/item/circuitboard/computer/cargo/express,
+/obj/item/circuitboard/computer/cargo,
 /obj/structure/closet/crate/engineering,
 /turf/open/floor/plasteel,
 /area/ship/cargo)
