+++ _maps/shuttles/independent/independent_junker.dmm
@@ -2277,7 +2277,7 @@
 	},
 /obj/effect/decal/cleanable/glass,
 /obj/structure/safe/floor,
-/obj/item/circuitboard/computer/cargo/express,
+/obj/item/circuitboard/computer/cargo,
 /turf/open/floor/pod/dark,
 /area/ship/crew/office)
 "Rj" = (
