+++ _maps/shuttles/independent/independent_kilo.dmm
@@ -1252,7 +1252,7 @@
 /area/ship/crew)
 "sW" = (
 /obj/effect/turf_decal/industrial/outline/yellow,
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/machinery/airalarm/directional/east,
