+++ _maps/shuttles/independent/independent_lagoon.dmm
@@ -256,7 +256,7 @@
 /turf/open/floor/plasteel,
 /area/ship/external)
 "bt" = (
-/obj/machinery/computer/cargo/express/retro{
+/obj/machinery/computer/cargo/retro{
 	dir = 4
 	},
 /turf/open/floor/plasteel/patterned/cargo_one,
