+++ _maps/shuttles/independent/independent_mudskipper.dmm
@@ -913,7 +913,7 @@
 /obj/structure/sign/warning/incident{
 	pixel_x = -32
 	},
-/obj/machinery/computer/cargo/express/retro{
+/obj/machinery/computer/cargo/retro{
 	dir = 4
 	},
 /turf/open/floor/plasteel/tech/grid,
