+++ _maps/shuttles/independent/independent_rigger.dmm
@@ -2231,7 +2231,7 @@
 /obj/machinery/atmospherics/components/unary/vent_pump/on/layer2{
 	dir = 4
 	},
-/obj/machinery/computer/cargo/express,
+/obj/machinery/computer/cargo,
 /obj/item/radio/intercom/directional/north,
 /turf/open/floor/carpet/blue,
 /area/ship/bridge)
