+++ _maps/shuttles/independent/independent_schmiedeberg.dmm
@@ -756,7 +756,7 @@
 /turf/open/floor/wood/walnut,
 /area/ship/crew/canteen)
 "ka" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 1
 	},
 /obj/item/radio/intercom/directional/south,
