+++ _maps/shuttles/independent/independent_shetland.dmm
@@ -5464,7 +5464,7 @@
 /obj/effect/turf_decal/corner/opaque/neutral/three_quarters{
 	dir = 1
 	},
-/obj/machinery/computer/cargo/express/retro,
+/obj/machinery/computer/cargo/retro,
 /turf/open/floor/plasteel/telecomms_floor,
 /area/ship/bridge)
 "VD" = (
