+++ _maps/shuttles/independent/independent_sunskipper.dmm
@@ -2050,7 +2050,7 @@
 /turf/open/floor/plasteel/tech,
 /area/ship/cargo)
 "AP" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	icon_state = "computer-right";
 	dir = 8
 	},
