+++ _maps/shuttles/inteq/inteq_colossus.dmm
@@ -852,7 +852,7 @@
 /area/ship/hallway/fore)
 "iX" = (
 /obj/item/radio/intercom/wideband/directional/south,
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 1
 	},
 /obj/effect/turf_decal/borderfloor,
