+++ _maps/shuttles/inteq/inteq_hound.dmm
@@ -1580,7 +1580,7 @@
 	dir = 1
 	},
 /obj/effect/turf_decal/steeldecal/steel_decals_central4,
-/obj/machinery/computer/cargo/express,
+/obj/machinery/computer/cargo,
 /obj/item/radio/intercom/wideband/directional/north{
 	pixel_y = 25;
 	pixel_x = 6
