+++ _maps/shuttles/inteq/inteq_talos.dmm
@@ -2329,7 +2329,7 @@
 /turf/open/floor/plasteel/tech,
 /area/ship/security/armory)
 "os" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/effect/turf_decal/corner/opaque/brown{
