+++ _maps/shuttles/inteq/inteq_valor.dmm
@@ -3855,7 +3855,7 @@
 /turf/open/floor/plasteel/patterned,
 /area/ship/cargo)
 "Ko" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 1
 	},
 /obj/effect/turf_decal/corner/opaque/brown{
