+++ _maps/shuttles/inteq/inteq_vaquero.dmm
@@ -2897,7 +2897,7 @@
 /turf/open/floor/plasteel/tech,
 /area/ship/cargo)
 "TK" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/effect/turf_decal/corner/opaque/yellow,
