+++ _maps/shuttles/nanotrasen/nanotrasen_delta.dmm
@@ -2530,7 +2530,7 @@
 /turf/open/floor/plasteel/dark,
 /area/ship/bridge)
 "Lv" = (
-/obj/machinery/computer/cargo/express/retro{
+/obj/machinery/computer/cargo/retro{
 	dir = 8
 	},
 /obj/effect/turf_decal/corner/opaque/nsorange,
