+++ _maps/shuttles/nanotrasen/nanotrasen_gecko.dmm
@@ -4283,7 +4283,7 @@
 /turf/open/floor/plasteel/grimy,
 /area/ship/crew)
 "RK" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 4
 	},
 /obj/item/radio/intercom/directional/west,
