+++ _maps/shuttles/nanotrasen/nanotrasen_heron.dmm
@@ -10099,7 +10099,7 @@
 /turf/open/floor/plasteel/tech,
 /area/ship/engineering)
 "LS" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/structure/railing{
