+++ _maps/shuttles/nanotrasen/nanotrasen_meta.dmm
@@ -1107,7 +1107,7 @@
 /turf/open/floor/plasteel/dark,
 /area/ship/cargo)
 "dg" = (
-/obj/machinery/computer/cargo/express/retro{
+/obj/machinery/computer/cargo/retro{
 	dir = 8
 	},
 /obj/effect/turf_decal/corner/opaque/blue/three_quarters{
@@ -2647,7 +2647,7 @@
 /area/ship/cargo)
 "Du" = (
 /obj/effect/turf_decal/corner/transparent/neutral/full,
-/obj/machinery/computer/cargo/express/retro{
+/obj/machinery/computer/cargo/retro{
 	dir = 8
 	},
 /obj/effect/decal/cleanable/dirt/dust,
