+++ _maps/shuttles/nanotrasen/nanotrasen_mimir.dmm
@@ -4024,7 +4024,7 @@
 /turf/open/floor/plasteel,
 /area/ship/security/prison)
 "xT" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /turf/open/floor/plasteel/telecomms_floor,
