+++ _maps/shuttles/nanotrasen/nanotrasen_osprey.dmm
@@ -95,7 +95,7 @@
 /obj/structure/disposalpipe/segment{
 	dir = 8
 	},
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/effect/turf_decal/corner/opaque/brown{
