+++ _maps/shuttles/nanotrasen/nanotrasen_ranger.dmm
@@ -464,7 +464,7 @@
 /turf/open/floor/plasteel/white,
 /area/ship/hallway/port)
 "eE" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 1
 	},
 /turf/open/floor/plasteel/tech,
