+++ _maps/shuttles/pgf/pgf_crying_sun.dmm
@@ -464,7 +464,7 @@
 /turf/open/floor/plasteel/telecomms_floor,
 /area/ship/hallway/port)
 "dK" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/structure/catwalk/over/plated_catwalk/dark,
