+++ _maps/shuttles/roumain/srm_elder.dmm
@@ -3609,7 +3609,7 @@
 /turf/open/floor/wood/maple,
 /area/ship/crew/cryo)
 "Sg" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/effect/turf_decal/spline/fancy/wood{
