+++ _maps/shuttles/solgov/solgov_chronicle.dmm
@@ -350,7 +350,7 @@
 /turf/open/floor/wood,
 /area/ship/crew)
 "da" = (
-/obj/machinery/computer/cargo/express/solgov{
+/obj/machinery/computer/cargo/solgov{
 	dir = 4
 	},
 /obj/item/radio/intercom/directional/west,
