+++ _maps/shuttles/solgov/solgov_inkwell.dmm
@@ -2139,7 +2139,7 @@
 /turf/open/floor/plasteel/mono,
 /area/ship/cargo)
 "nO" = (
-/obj/machinery/computer/cargo/express/solgov,
+/obj/machinery/computer/cargo/solgov,
 /turf/open/floor/wood/maple,
 /area/ship/bridge)
 "nR" = (
