+++ _maps/shuttles/solgov/solgov_paracelsus.dmm
@@ -2725,7 +2725,7 @@
 /obj/structure/railing/wood{
 	dir = 8
 	},
-/obj/machinery/computer/cargo/express/solgov{
+/obj/machinery/computer/cargo/solgov{
 	dir = 4
 	},
 /obj/effect/turf_decal/siding/wood{
