+++ _maps/shuttles/syndicate/syndicate_cybersun_kansatsu.dmm
@@ -258,7 +258,7 @@
 /turf/open/floor/plasteel/stairs,
 /area/ship/cargo)
 "fk" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /obj/machinery/button/door{
