+++ _maps/shuttles/syndicate/syndicate_gorlex_hyena.dmm
@@ -952,7 +952,7 @@
 "pI" = (
 /obj/item/radio/intercom/wideband/directional/east,
 /obj/effect/decal/cleanable/dirt/dust,
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 1
 	},
 /obj/effect/turf_decal/borderfloorblack,
