+++ _maps/shuttles/syndicate/syndicate_gorlex_komodo.dmm
@@ -3507,7 +3507,7 @@
 /turf/open/floor/mineral/plastitanium/red,
 /area/ship/hallway/central)
 "HI" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	layer = 3
 	},
 /obj/effect/turf_decal/techfloor{
