+++ _maps/shuttles/syndicate/syndicate_litieguai.dmm
@@ -2608,7 +2608,7 @@
 /turf/open/floor/plasteel/white,
 /area/ship/hallway/central)
 "TA" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /turf/open/floor/plasteel/dark,
