+++ _maps/shuttles/syndicate/syndicate_panacea.dmm
@@ -887,7 +887,7 @@
 /turf/open/floor/suns/dark/plain,
 /area/ship/crew/office/lobby)
 "fG" = (
-/obj/machinery/computer/cargo/express{
+/obj/machinery/computer/cargo{
 	dir = 8
 	},
 /turf/open/floor/suns/dark,
