+++ check_regex.yaml
@@ -31,7 +31,7 @@ standards:
     - exactly: [4, "/mob text paths", '"/mob']
     - exactly: [43, "/obj text paths", '"/obj']
     - exactly: [0, "/turf text paths", '"/turf']
-    - exactly: [117, "text2path uses", "text2path"]
+    - exactly: [116, "text2path uses", "text2path"]
 
     - exactly: [18, "world<< uses", 'world[ \t]*<<']
     - exactly: [0, "world.log<< uses", 'world.log[ \t]*<<']
