+++ shiptest.dme
@@ -1885,9 +1885,9 @@
 #include "code\modules\cargo\bounty.dm"
 #include "code\modules\cargo\bounty_console.dm"
 #include "code\modules\cargo\centcom_podlauncher.dm"
+#include "code\modules\cargo\console.dm"
 #include "code\modules\cargo\export_scanner.dm"
 #include "code\modules\cargo\exports.dm"
-#include "code\modules\cargo\expressconsole.dm"
 #include "code\modules\cargo\gondolapod.dm"
 #include "code\modules\cargo\order.dm"
 #include "code\modules\cargo\packs.dm"
